all errors fix except reactrouter dom
