--- app/views/root/index.html.erb
@@ -1,1 +1,0 @@
-<div id="root">
